requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap4==2.1.1
django-extensions==2.2.9
django-qr-code==1.2.0
djangorestframework==3.11.0
firebase-admin==4.3.0
fpdf==1.7.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
Pillow==7.2.0
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.5.3
pypng==0.0.20
PyQRCode==1.2.1
pytz==2020.1
qrcode==6.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
